Fixed argument type

+++ source/pybase.cpp
@@ -483,7 +483,7 @@ static bool getmodulesub(const char *mod,char *dir,int len,const char *ext)
     return name != NULL;
 #elif FLEXT_SYS == FLEXT_SYS_MAX
     short path;
-    long type;
+    unsigned int type;
     char smod[1024];
     strcpy(smod,mod);
     strcat(smod,ext);
